Updated db structure

+++ db_setup.sql
@@ -10,7 +10,7 @@ CREATE TABLE users(
     lastName    varchar(25) not null,
     username    varchar(40) unique,
     picture_url varchar(500),
-    tracking_id int unique not null,
+    tracking_id bigint unique not null,
     tracking_token varchar(256) not null,
     organization_id int
 );
@@ -25,7 +25,7 @@ CREATE TABLE activities(
     longitude   int not null,
     started_at  timestamp not null,
     ended_at    timestamp not null,
-    tracking_id int unique not null,
+    tracking_id bigint unique not null,
     user_id     int not null,
     FOREIGN key (user_id)
         REFERENCES users (id)
